+++ extensions/linear/src/components/docs/DocumentList.tsx
@@ -5,6 +5,7 @@ import { useDocuments } from "../../hooks/useDocuments";
 import { getProjectIcon } from "../../helpers/projects";
 import { Document } from "./Document";
 import { ProjectResult } from "../../api/getProjects";
+import { useInitiatives } from "../../hooks/useInitiatives";
 
 type DocumentListProps = {
   project?: Pick<ProjectResult, "id" | "name">;
@@ -15,6 +16,7 @@ export function DocumentList({ project }: DocumentListProps) {
   const [projectId, setProjectId] = useState<string>("");
 
   const { projects, isLoadingProjects } = useProjects();
+  const { initiatives, isLoadingInitiatives } = useInitiatives();
   const { docs, isLoadingDocs, supportsDocTypeahead, mutateDocs } = useDocuments(query, projectId);
 
   const filteredDocs = useMemo(() => {
@@ -27,25 +29,31 @@ export function DocumentList({ project }: DocumentListProps) {
       return docs;
     }
 
-    return docs?.filter((doc) => doc.project.id === id) || [];
+    return docs?.filter((doc) => (doc.project?.id ?? "") === id) || [];
   }, [project, projectId, docs, projects]);
 
   return (
     <List
-      isLoading={isLoadingProjects || isLoadingDocs}
+      isLoading={isLoadingProjects || isLoadingDocs || isLoadingInitiatives}
       navigationTitle={project ? `${project.name} Documents` : undefined}
-      {...(!project && projects && projects.length > 0
+      {...(!project && (projects ?? []).length > 0
         ? {
             searchBarAccessory: (
-              <List.Dropdown tooltip="Change Project" placeholder="Search Project" onChange={setProjectId} storeValue>
+              <List.Dropdown
+                tooltip="Change Parent"
+                onChange={setProjectId}
+                storeValue
+                filtering={{ keepSectionOrder: true }}
+              >
                 <List.Dropdown.Item value="" title="All Documents" />
 
-                <List.Dropdown.Section>
+                <List.Dropdown.Section title="Projects">
                   {projects?.map((project) => (
                     <List.Dropdown.Item
                       key={project.id}
                       value={project.id}
                       title={project.name}
+                      keywords={[project.name, project.description]}
                       icon={getProjectIcon(project)}
                     />
                   ))}
@@ -59,7 +67,7 @@ export function DocumentList({ project }: DocumentListProps) {
         : { onSearchTextChange: setQuery, searchBarPlaceholder: "Search by document title", throttle: true })}
     >
       {filteredDocs.map((doc) => (
-        <Document key={doc.id} doc={doc} mutateDocs={mutateDocs} />
+        <Document key={doc.id} {...{ doc, mutateDocs, projects, initiatives }} />
       ))}
       <List.EmptyView title="No documents found" icon={{ source: Icon.Document, tintColor: Color.Orange }} />
     </List>